cloud/static-analyzer/requirements.txt
# AegisAI Static Analyzer Requirements

# Core Dependencies
pefile>=2021.5.24
yara-python>=4.0.0
requests>=2.25.0
aiohttp>=3.7.0

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Utilities
pydantic>=1.8.0
python-dateutil>=2.8.0